% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_CI.R
\name{bayes_CI}
\alias{bayes_CI}
\title{Creation of Bayesian Credible Interval from MH Samples}
\usage{
bayes_CI(mh_output, alpha)
}
\arguments{
\item{mh_output}{- vector output from met_hastings function}

\item{alpha}{- value of bounds, calculated as alpha, 1-alpha/2}
}
\value{
Lower and upper bounds of the alpha interval
}
\description{
Creation of Bayesian Credible Interval from MH Samples
}
\examples{
  
95\% credible interval for our draws in example #1

out <- met_hastings(10000, start = 1, burn_in = 1000,
  jparams = 2,distr = "gamma",c(1.7,4.4))
bayes_CI(out, 0.1)
}
