% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distr.R
\name{draw_jump}
\alias{draw_jump}
\title{Functions generating draws and calculating the acceptance ratio}
\usage{
To structure your jumping or distribution parameters, look at how the function 
e.g. for a normal distribution, jparams = (your standard deviation)
If using a custom structure, the likelihood parameters should be passed to dparams
and the likelihood function should include input for theta
See example #3 in the main function for assisstence in structuring your likelihood function
}
\description{
Functions generating draws and calculating the acceptance ratio
}
