% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_hastings.R
\name{met_hastings}
\alias{met_hastings}
\title{Metropolis-Hastings Algorithmn (non-adaptive)
Random Walk}
\usage{
met_hastings(
  nsims,
  start = 1,
  burn_in,
  jump = "normal",
  jparams = 1,
  distr = "normal",
  dparams = 1,
  likelihood = NULL
)
}
\description{
Metropolis-Hastings Algorithmn (non-adaptive)
Random Walk
}
